Python/restart_pads/requirements.txt
psycopg2-binary
python-nmap
ipaddress
configparser
netmiko